+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/ListFileExample.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.ftp.example;
+package com.dwarfeng.ftp.node.example;
 
 import com.dwarfeng.ftp.stack.bean.dto.FtpFile;
 import com.dwarfeng.ftp.stack.handler.FtpHandler;
@@ -9,12 +9,12 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
 import java.util.Scanner;
 
 /**
- * 描述文件示例。
+ * 列出文件示例。
  *
  * @author DwArFeng
- * @since 1.2.0
+ * @since 1.1.7
  */
-public class DescFileExample {
+public class ListFileExample {
 
     public static void main(String[] args) throws Exception {
         ClassPathXmlApplicationContext ctx = new ClassPathXmlApplicationContext(
@@ -29,7 +29,7 @@ public class DescFileExample {
 
         // 显示欢迎信息并获取展示功能用的根文件夹。
         System.out.println("开发者您好!");
-        System.out.println("这是一个示例, 用于演示 dwarfeng-ftp 的描述文件功能");
+        System.out.println("这是一个示例, 用于演示 dwarfeng-ftp 的列出文件功能");
         System.out.println("该示例将会在你配置的 ftp 目录下新建名为 foobar 的文件夹, 如果您的 ftp 已经有这个文件夹了, " +
                 "请指定一个不存在的文件夹");
         System.out.print("请指定一个文件夹用于演示功能, 不填默认为 foobar...");
@@ -42,44 +42,39 @@ public class DescFileExample {
         byte[] content = ResourceUtil.getContent();
         System.out.println();
         System.out.println("1. 创建演示文件...");
-        ftpHandler.storeFile(new String[]{rootPath, "comic-girl"}, "comic-girl.jpg", content);
-        ftpHandler.storeFile(new String[]{rootPath, "漫画女孩"}, "漫画女孩.jpg", content);
+        ftpHandler.storeFile(new String[]{rootPath}, "comic-girl.jpg", content);
+        ftpHandler.storeFile(new String[]{rootPath}, "漫画女孩.jpg", content);
         System.out.println("文件新建完毕, 您将会看到文件被创建");
         System.out.print("请按回车键继续...");
         scanner.nextLine();
 
-        // 2. 描述存在的文件及目录。
+        // 2. 列出文件。
         System.out.println();
-        System.out.println("2. 描述存在的文件及目录...");
-        FtpFile ftpFile;
-        ftpFile = ftpHandler.descFile(new String[]{rootPath}, "comic-girl");
-        System.out.printf("/%s/comic-girl: %s\n", rootPath, ftpFile);
-        ftpFile = ftpHandler.descFile(new String[]{rootPath, "comic-girl"}, "comic-girl.jpg");
-        System.out.printf("/%s/comic-girl/comic-girl.jpg: %s\n", rootPath, ftpFile);
-        ftpFile = ftpHandler.descFile(new String[]{rootPath}, "漫画女孩");
-        System.out.printf("/%s/漫画女孩: %s\n", rootPath, ftpFile);
-        ftpFile = ftpHandler.descFile(new String[]{rootPath, "漫画女孩"}, "漫画女孩.jpg");
-        System.out.printf("/%s/漫画女孩/漫画女孩.jpg: %s\n", rootPath, ftpFile);
+        System.out.println("2. 列出文件...");
+        FtpFile[] ftpFiles = ftpHandler.listFiles(new String[]{rootPath});
+        System.out.println("列出文件夹 " + rootPath + " 下的文件: ");
+        for (FtpFile ftpFile : ftpFiles) {
+            System.out.println("  " + ftpFile.toString());
+        }
         System.out.print("请按回车键继续...");
         scanner.nextLine();
 
-        // 3. 描述不存在的文件及目录。
+        // 3. 列出文件名。
         System.out.println();
-        System.out.println("3. 描述不存在的文件及目录...");
-        ftpFile = ftpHandler.descFile(new String[]{rootPath}, "何もありません");
-        System.out.printf("/%s/何もありません: %s\n", rootPath, ftpFile);
-        ftpFile = ftpHandler.descFile(new String[]{rootPath}, "何もありません.jpg");
-        System.out.printf("/%s/何もありません.jpg: %s\n", rootPath, ftpFile);
+        System.out.println("3. 列出文件名...");
+        String[] fileNames = ftpHandler.listFileNames(new String[]{rootPath});
+        System.out.println("列出文件夹 " + rootPath + " 下的文件名: ");
+        for (String fileName : fileNames) {
+            System.out.println("  " + fileName);
+        }
         System.out.print("请按回车键继续...");
         scanner.nextLine();
 
         // 4. 文件删除。
         System.out.println();
         System.out.println("4. 文件删除...");
-        ftpHandler.deleteFile(new String[]{rootPath, "comic-girl"}, "comic-girl.jpg");
-        ftpHandler.deleteFile(new String[]{rootPath, "漫画女孩"}, "漫画女孩.jpg");
-        ftpHandler.removeDirectory(new String[]{rootPath, "comic-girl"});
-        ftpHandler.removeDirectory(new String[]{rootPath, "漫画女孩"});
+        ftpHandler.deleteFile(new String[]{rootPath}, "comic-girl.jpg");
+        ftpHandler.deleteFile(new String[]{rootPath}, "漫画女孩.jpg");
         System.out.println("示例演示完毕, 感谢您测试与使用!");
 
         ctx.stop();