+++ dwarfeng-ftp-core/src/test/java/com/dwarfeng/ftp/node/example/StreamExample.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.ftp.example;
+package com.dwarfeng.ftp.node.example;
 
 import com.dwarfeng.dutil.basic.io.IOUtil;
 import com.dwarfeng.ftp.stack.handler.FtpHandler;